backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.35
email-validator>=2.0.0
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic==2.9.2
pydantic-settings==2.5.2
google-auth[requests]==2.35.0
python-multipart==0.0.9

